\name{bentcableAR-internal}
\alias{alph.vect}
\alias{alpha}
\alias{ar.p.resid.core}
\alias{basic.cable}
\alias{cable.Sig.ar.p}
\alias{cable.design.mat}
\alias{cable.diagnose.resid}
\alias{cable.loop}
\alias{cable.loop.ar.p}
\alias{find.fisher}
\alias{grad}
\alias{loglik.given.g}
\alias{loglik.given.g.ar.p}
\alias{loglik.surf}
\alias{loglik.surf.ar.p}
\alias{pick.peak}
\alias{sse.ar.p}
\alias{sse.ar.p.core}
\alias{sse.ar.p.given.beta.phi}
\alias{sse.ar.p.given.g}
\alias{sse.ar.p.given.t.g}
\title{Internal 'bentcableAR' Functions}
\description{
  Internal functions for the \code{bentcableAR} package.
}
\details{
  These are not intended to be called by the user.
}
\keyword{ internal }

