\name{stagnant}
\docType{data}
\alias{stagnant}
\title{ Stagnant Band Height Data }
\description{
	This is the well-studied data from R.A. Cook's Ph.D.
	thesis (Queen's University) on the relationship between the flow
	rate and band height of water on an incline.
}
\usage{data(stagnant)}
\format{A data frame with two columns:
	\itemize{
		\item{\code{logflow}}{ The flow rate in g/cm-s, log-transformed }
		\item{\code{loght}}{ The band height in cm, log-transformed. }
	}
}
\source{ Seber, G. A. F. and Wild, C. J. (2003), \emph{Nonlinear
Regression}, Hoboken: Wiley. }
\references{
Chiu, G., Lockhart, R. and Routledge, R. (2006), Bent-Cable Regression
	Theory and Applications, \emph{Journal of the American Statistical
	Association}, \bold{101}, 542--553.
	\url{http://faculty.washington.edu/gchiu/Articles/bentcable-jasa.pdf}
}
\keyword{datasets}
